requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
peewee==3.14.4
psycopg2==2.8.6
pycparser==2.20
python-dotenv==0.17.1
six==1.16.0
virtualenv==20.4.6
Werkzeug==1.0.1
